mainprocess/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
dnspython==2.4.2
minio==7.2.3
pycparser==2.21
pycryptodome==3.20.0
pymongo==4.6.1
python-dotenv==1.0.0
typing_extensions==4.9.0
urllib3==2.1.0
